Update sound card name on Rock 5B
The analog sound card is currently registered in the system under the
'Analog' name, which is too generic and was not accepted by the
alsa-project/alsa-ucm-conf maintainer.

The Rock 5B DTS has been updated so that the audio-graph-card driver
will use the more specific naming 'rk3588-es8316', which is still
generic enough to be shared with other compatible boards, e.g. Rock 5A.
The audio card will be listed as:

  card 0: rk3588es8316 [rk3588-es8316]

The kernel patch has been already merged upstream. This patch only
handles the necessary UCM configuration changes - just renaming the
'conf.d' entries.

Signed-off-by: Cristian Ciocaltea <cristian.ciocaltea@collabora.com>