Update u-boot branch to rk3588
As we want to have a more generic name for the u-boot branch for the
rk3588 development, we changed the branch name to just `rk3588` as it's
not just about the rock-5b board but about the rk3588 SoC and associated
boards.
Hence, rename the branch used for the debian image.

Signed-off-by: Eugen Hristev <eugen.hristev@collabora.com>