README: give some more user documentation
The current README is a mix of flash and build instructions, but no
usage instructions. To save users the trouble of looking through
the debos recipe to find out what the username and password combo
is, and stop them from getting confused and chasing down rabbit
holes they don't need to chase down if they just want to use the
images, more clearly separate the README into sections and give
user information.

Signed-off-by: Nicolas Frattaroli <nicolas.frattaroli@collabora.com>