Clean rootfs of various files that should be per-device
During image generation, a few files happen to be created as
apt installs packages and such. This includes the SSH host keys,
the machine-id and random seeds. These things should be unique and
not carried over across multiple deployments of the same image.

Thus, we add a new script to clean the rootfs, and a new systemd
service to regenerate the SSH host keys if they're missing.

Signed-off-by: Nicolas Frattaroli <nicolas.frattaroli@collabora.com>