Enable mDNS (Multicast DNS) announcements
The images already use systemd-networkd and systemd-resolved. The
latter, by default, will announce itself over the mDNS protocol if
the network link it is connected to has Multicast and MulticastDNS
enabled.

If two devices with the same hostname announce themselves on the
same network, the conflict is resolved by appending an incrementing
number at the end of the hostname.

Enable this feature by setting MulticastDNS in the systemd-networkd
config of the wired link, as well as the Multicast setting. The
boards can now be reached at their hostname + ".local" on any system
that has mDNS resolution enabled, e.g. Windows and macOS by default.

Signed-off-by: Nicolas Frattaroli <nicolas.frattaroli@collabora.com>