Don't write "uninitialized" to machine-id
Having a /etc/machine-id with `uninitialized\n` as the contents
triggers systemd's ConditionFirstBoot. In our images, this will
start `systemd-firstboot.service`, which may interactively prompt
for locale/keymap/timezone/password options.

While I have not observed this happening on a serial terminal, this is
certainly not what we want. Instead make /etc/machine-id an empty file
as per sjoerd's suggestion (and the machine-id manpage) to allow systemd
to still bind-mount onto the file and generate a new machine-id, but not
treat it as the first boot.

Signed-off-by: Nicolas Frattaroli <nicolas.frattaroli@collabora.com>