Add recipe for Sige5 images
Add a recipe for RK3576 boards. The utilised u-boot is based on a
mainline-based branch by Kwiboo.  As mainline U-Boot packs both TPL and
SPL into one binary, we write just `u-boot-rockchip.bin` at the start
of the image where the TPL usually resides.

The utilised BL31 is from rkbin (proprietary), along with the TPL.

As it stands, the images hang on boot because the included kernel has a
regression with regards to eMMC. However, disabling the eMMC in that
kernel's DT does let the board boot fully. If no quick fix for the eMMC
regression is found, that is one option.

We also adjust the README and gitlab CI to be in line with the new
supported target's existence, and rename and adjust the artifact
downloading script.

Fixes #1.

Signed-off-by: Nicolas Frattaroli <nicolas.frattaroli@collabora.com>