Parallelize image builds
Reduce total build time by doing more parallelization. The images
for the CI is more or less independent from the normal images and
just shares some definitions. Also combining the rootfs with the
bootloader to create the board specific images can be run for all
boards in parallel once the rootfs has been generated. This should
speed up the CI pipeline a bit.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>