Drop extra CI initramfs build
The CI initramfs build is not using mkinitramfs and thus not
functional for LAVA, since it does not support booting from
NFS without kernel support. The origin of this extra image is
the CI healthcheck image, which we used previously for our
LAVA boot tests. Apparently I should never have copied this
part.

It has not been noticed so far, since we also build a working
initramfs from the ci-rootfs.yaml. As the broken one is unused
we can just drop it and afterwards simplify the CI image
definition to a single file.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>